+++ zenburn-theme.el
@@ -229,6 +229,8 @@ Also bind `class' to ((class color) (min-colors 89))."
    `(custom-state ((t (:foreground ,zenburn-green+4))))
 ;;;;; display-fill-column-indicator
    `(fill-column-indicator ((,class :foreground ,zenburn-bg-05 :weight semilight)))
+;;;;; eldoc
+   `(eldoc-highlight-function-argument ((t (:foreground ,zenburn-yellow :weight bold))))
 ;;;;; eww
    '(eww-invalid-certificate ((t (:inherit error))))
    '(eww-valid-certificate   ((t (:inherit success))))