+++ zenburn-theme.el
@@ -330,6 +330,9 @@ Also bind `class' to ((class color) (min-colors 89))."
    `(newsticker-treeview-obsolete-face ((t (:foreground ,zenburn-red))))
    `(newsticker-treeview-old-face ((t (:foreground ,zenburn-bg+3))))
    `(newsticker-treeview-selection-face ((t (:background ,zenburn-bg-1 :foreground ,zenburn-yellow))))
+;;;;; pulse
+   `(pulse-highlight-face ((t (:background ,zenburn-blue-5))))
+   `(pulse-highlight-start-face ((t (:background ,zenburn-blue-5))))
 ;;;;; woman
    '(woman-bold   ((t (:inherit font-lock-keyword-face))))
    '(woman-italic ((t (:inherit (font-lock-string-face italic)))))