+++ zenburn-theme.el
@@ -1146,6 +1146,12 @@ the just-introduced bindings."
    `(mm-uu-extract ((t (:background ,zenburn-bg-05 :foreground ,zenburn-green+1))))
 ;;;;; go-guru
    `(go-guru-hl-identifier-face ((t (:foreground ,zenburn-bg-1 :background ,zenburn-green+1))))
+;;;;; gptel
+   `(gptel-context-highlight-face ((t (:background ,zenburn-bg+05 :extend t))))
+   `(gptel-context-deletion-face ((t (:background ,zenburn-diff-removed-bg :extend t))))
+   `(gptel-rewrite-highlight-face ((t (:background ,zenburn-diff-changed-bg :extend t))))
+   `(gptel-response-highlight ((t (:background ,zenburn-bg-05 :extend t))))
+   `(gptel-response-fringe-highlight ((t (:foreground ,zenburn-blue))))
 ;;;;; hackernews
    '(hackernews-comment-count ((t (:inherit link-visited :underline nil))))
    '(hackernews-link          ((t (:inherit link         :underline nil))))