+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `lsp-mode` and `lsp-ui` support.
 * Add `hl-todo` support.
 * Add `nerd-icons` support.
+* Add `helpful` support.
 
 ### Changes