+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add `wgrep` support.
 * Add `multiple-cursors` support.
 * Add `eldoc` support (`eldoc-highlight-function-argument`).
+* Add `pulse` support.
 
 ### Changes