+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add `multiple-cursors` support.
 * Add `eldoc` support (`eldoc-highlight-function-argument`).
 * Add `pulse` support.
+* Add `treemacs` support.
 
 ### Changes